// frontend/src/shared/lib/api/model/protocolFormConditionTemplate.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Cellar
 * OpenAPI spec version: 1.3.0-17-gd9da37bb-dirty
 */
import type { ProtocolFormConditionTemplateUnit } from './protocolFormConditionTemplateUnit';
import type { ProtocolFormConditionTemplatePickListValues } from './protocolFormConditionTemplatePickListValues';

export interface ProtocolFormConditionTemplate {
  name: string;
  data_type: string;
  unit?: ProtocolFormConditionTemplateUnit;
  pick_list_values?: ProtocolFormConditionTemplatePickListValues;
}