// frontend/src/shared/lib/api/model/protocolFormReadoutTemplate.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Cellar
 * OpenAPI spec version: 1.3.0-17-gd9da37bb-dirty
 */
import type { ProtocolFormReadoutTemplateUnit } from './protocolFormReadoutTemplateUnit';
import type { ProtocolFormReadoutTemplateCalculationFormula } from './protocolFormReadoutTemplateCalculationFormula';
import type { ProtocolFormReadoutTemplatePickListValues } from './protocolFormReadoutTemplatePickListValues';
import type { ProtocolFormReadoutTemplateDoseResponseConfig } from './protocolFormReadoutTemplateDoseResponseConfig';

export interface ProtocolFormReadoutTemplate {
  name: string;
  data_type: string;
  unit?: ProtocolFormReadoutTemplateUnit;
  aggregation?: string;
  normalization?: string;
  is_calculated?: boolean;
  calculation_formula?: ProtocolFormReadoutTemplateCalculationFormula;
  pick_list_values?: ProtocolFormReadoutTemplatePickListValues;
  dose_response_config?: ProtocolFormReadoutTemplateDoseResponseConfig;
}